modularize eth_wallet
