+++ crates/madara/src/config/mod.rs
@@ -3,4 +3,7 @@ pub mod constants;
 pub mod global_config;
 pub mod madara;
 
+mod eth_wallet;
 mod l1_config;
+
+pub use eth_wallet::EthWallet;
